Implement email sender command to creditMemoGeneretedEventListener
