+++ tests/Application/config/packages/test/knp_snappy.yaml
@@ -1,3 +1,3 @@
-#knp_snappy:
-#    pdf:
-#        binary: "%kernel.project_dir%/etc/wkhtmltopdf"
+knp_snappy:
+    pdf:
+        binary: "%kernel.project_dir%/etc/wkhtmltopdf"
